borrowck: more eagerly prepopulate opaques
